Refactored into two modules.
