--- fcrepo-import-export-common/.gitignore
@@ -1,1 +1,0 @@
-/target/
