--- fcrepo-import-export-driver/.gitignore
@@ -1,1 +1,0 @@
-/target/
