+++ fcrepo-import-export-driver/src/test/resources/sample/binary/rdf/rest/bin1/fcr_metadata.jsonld
