+++ itests/src/test/java/org/fcrepo/importexport/driver/integration/AbstractResourceIT.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed to DuraSpace under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * DuraSpace licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except in
+ * compliance with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.fcrepo.importexport.driver.integration;
+
+import java.util.concurrent.TimeUnit;
+
+import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
+import org.fcrepo.client.FcrepoClient;
+
+/**
+ * @author awoods
+ * @since 2016-09-01
+ */
+abstract class AbstractResourceIT {
+
+    private static final int SERVER_PORT = Integer.parseInt(System.getProperty("fcrepo.dynamic.test.port", "8080"));
+
+    private static final String HOSTNAME = "localhost";
+
+    static final String serverAddress = "http://" + HOSTNAME + ":" + SERVER_PORT + "/rest/";
+
+    static FcrepoClient client;
+
+    AbstractResourceIT() {
+        final PoolingHttpClientConnectionManager connectionManager = new PoolingHttpClientConnectionManager();
+        connectionManager.setMaxTotal(Integer.MAX_VALUE);
+        connectionManager.setDefaultMaxPerRoute(20);
+        connectionManager.closeIdleConnections(3, TimeUnit.SECONDS);
+    }
+}
