+++ itests/src/test/resources/logback-test.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE configuration>
+<configuration>
+  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+    <encoder>
+      <pattern>%p %d{HH:mm:ss.SSS} \(%c{0}\) %m%n</pattern>
+    </encoder>
+  </appender>
+
+  <logger name="org.fcrepo.importexport" additivity="false" level="${fcrepo.importexport.log:-DEBUG}">
+    <appender-ref ref="STDOUT"/>
+  </logger>
+  <logger name="org.fcrepo.exporter" additivity="false" level="${fcrepo.exporter.log:-DEBUG}">
+    <appender-ref ref="STDOUT"/>
+  </logger>
+  <logger name="org.fcrepo" additivity="false" level="${fcrepo.log:-INFO}">
+    <appender-ref ref="STDOUT"/>
+  </logger>
+  <logger name="org.glassfish" additivity="false" level="ERROR">
+    <appender-ref ref="STDOUT"/>
+  </logger>
+  <root additivity="false" level="WARN">
+    <appender-ref ref="STDOUT"/>
+  </root>
+</configuration>
