+++ itests/src/test/resources/spring-test/rest.xml
