+++ pom.xml
@@ -44,10 +44,8 @@
   </properties>
 
   <modules>
-    <module>fcrepo-import</module>
-    <module>fcrepo-export</module>
-    <module>fcrepo-import-export-common</module>
     <module>fcrepo-import-export-driver</module>
+    <module>itests</module>
   </modules>
 
   <dependencyManagement>
@@ -73,19 +71,7 @@
 
       <dependency>
         <groupId>org.fcrepo.importexport</groupId>
-        <artifactId>fcrepo-import-export-common</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.fcrepo.importexport</groupId>
-        <artifactId>fcrepo-import</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.fcrepo.importexport</groupId>
-        <artifactId>fcrepo-export</artifactId>
+        <artifactId>fcrepo-import-export-driver</artifactId>
         <version>${project.version}</version>
       </dependency>
